fixed: #82
A bypass counter is added to the object getter and create methods.
This bypass counter is used to bypass the cache if it is bigger than one.
For each bypass it will be decreased by one. If the counter reaches zero,
then the cache mechanism runs as previously.
The getMetaObjectArray method sets the bypass counter to one, so the
dummy object can avoid to be cached, and all objects below it will be
cached.