[issue #86] DialectProvider and several other hacks
The DialectProvider is introduced to properly propagate the dialect.
The only reason for this is the fact that actually the client builds
bean representations of objects and thus needs some dialect dependent
operations. Additionally, the bean factory contains some temporary hacks
to actually make it work in oracle mode. These issues have to be solved
properly for producation.