issue #98
new dependency  => quartz

ServerActions are scheduled if they implement the ScheduledServerAction
interface. Special ServerActionParameter Keys:
SSAPK_DELAY : date marking the beginn of the exection (first run)
SSAPK_RULE : cron rule for the exectution
SSAPK_ABORT : for abortion of a ScheduledServerAction

ScheduledServerAction "processes" are identified by their key.
The invokation of a new ScheduledServerActions  with an key
that are already used by another ScheduledServerAction aborts the existing
ScheduledServerAction an invokes the new on.